voucher price constants on chain client side

+++ client/src/actions/init_neighborhood_metadata.ts
@@ -28,7 +28,7 @@ export class InitNeighborhoodMetadataInstructionData {
           ["price", "u64"],
           ["voucherLiveDate", "u64"],
           ["voucherReceiveLimit", "u64"],
-          ["voucherPriceCoefficient", "f64"],
+          ["voucherPriceCoefficient", "u64"],
         ],
       },
     ],
