+++ space/src/instruction.rs
@@ -16,7 +16,7 @@ pub struct InitNeighborhoodMetadataArgs {
     pub neighborhood_name: [u8; 64],
     pub voucher_live_date: u64,
     pub voucher_receive_limit: u64,
-    pub voucher_price_coefficient: f64,
+    pub voucher_price_coefficient: u64,
 }
 
 #[repr(C)]
